Define Doctype at render time.
Added constants to Documont class for all the doctypes as per http://www.w3.org/QA/2002/04/valid-dtd-list.html which includes SVG MathML and combinations.
Referenced the Document constants for HTML documents in the Html class for convenience.
Changed the Document render method to render XML instead of HTMLand prepend the doctype if supplied.